Resource based authorization

+++ Restaurants.Application/Dishes/Commands/DeleteDishes/DeleteDishesForRestaurantCommandHandler.cs
@@ -1,7 +1,9 @@
 ﻿using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repsitories;
 
 namespace Restaurants.Application.Dishes.Commands.DeleteDishes;
@@ -11,14 +13,17 @@ public class DeleteDishesForRestaurantCommandHandler : IRequestHandler<DeleteDis
     private readonly ILogger<DeleteDishesForRestaurantCommandHandler> _logger;
     private readonly IRestaurantsRepository _restaurantsRepository;
     private readonly IDishiesRepository _dishiesRepository;
+    private readonly IRestaurantAuthorizationService _restaurantAuthorizationService;
 
     public DeleteDishesForRestaurantCommandHandler(ILogger<DeleteDishesForRestaurantCommandHandler> logger ,
                                                    IRestaurantsRepository restaurantsRepository ,
-                                                   IDishiesRepository dishiesRepository)
+                                                   IDishiesRepository dishiesRepository ,
+                                                   IRestaurantAuthorizationService restaurantAuthorizationService)
     {
         _logger = logger;
         _restaurantsRepository = restaurantsRepository;
         _dishiesRepository = dishiesRepository;
+        _restaurantAuthorizationService = restaurantAuthorizationService;
     }
 
     public async Task Handle(DeleteDishesForRestaurantCommand request , CancellationToken cancellationToken)
@@ -27,6 +32,9 @@ public class DeleteDishesForRestaurantCommandHandler : IRequestHandler<DeleteDis
 
         var restaurant = await _restaurantsRepository.GetByIdAsync(request.RestaurantId)
             ?? throw new NotFoundException(nameof(Restaurant) , request.RestaurantId.ToString());
+
+        if (!_restaurantAuthorizationService.Authorize(restaurant , ResourceOperation.Delete))
+            throw new ForbidException();
 
         await _dishiesRepository.Delete(restaurant.Dishes);
     }
