+++ Restaurants.Application/Restaurants/Commands/DeleteRestaurant/DeleteRestaurantCommandHandler.cs
@@ -1,7 +1,9 @@
 ﻿using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repsitories;
 
 namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant;
@@ -10,12 +12,15 @@ public class DeleteRestaurantCommandHandler : IRequestHandler<DeleteRestaurantCo
 {
     private readonly ILogger<DeleteRestaurantCommandHandler> _logger;
     private readonly IRestaurantsRepository _restaurantsRepository;
+    private readonly IRestaurantAuthorizationService _restaurantAuthorizationService;
 
     public DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandler> logger ,
-                                          IRestaurantsRepository restaurantsRepository)
+                                          IRestaurantsRepository restaurantsRepository ,
+                                          IRestaurantAuthorizationService restaurantAuthorizationService)
     {
         _logger = logger;
         _restaurantsRepository = restaurantsRepository;
+        _restaurantAuthorizationService = restaurantAuthorizationService;
     }
 
     public async Task Handle(DeleteRestaurantCommand request , CancellationToken cancellationToken)
@@ -23,6 +28,9 @@ public class DeleteRestaurantCommandHandler : IRequestHandler<DeleteRestaurantCo
         _logger.LogInformation("Delete restaurant by Id: {RestaurantId}" , request.Id);
         var restaurants = await _restaurantsRepository.GetByIdAsync(request.Id)
             ?? throw new NotFoundException(nameof(Restaurant) , request.Id.ToString());
+
+        if (!_restaurantAuthorizationService.Authorize(restaurants, ResourceOperation.Delete))
+            throw new ForbidException();
 
         await _restaurantsRepository.DeleteRestaurant(restaurants);
     }
